[CS-1742] - Organize files into folders w/ context
