+++ cardstack/src/components/TransactionConfirmationSheet/displays/PrepaidCard/TransferPrepaidCardDisplay.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { ContactAvatar } from '../../../../src/components/contacts';
-import { TransactionConfirmationSectionHeaderText } from './TransactionConfirmationSectionHeaderText';
+import { ContactAvatar } from '../../../../../../src/components/contacts';
+import { SectionHeaderText } from '../components/SectionHeaderText';
 import {
   Container,
   HorizontalDivider,
@@ -53,9 +53,7 @@ const FromSection = () => {
 
   return (
     <Container marginTop={8} width="100%">
-      <TransactionConfirmationSectionHeaderText>
-        FROM
-      </TransactionConfirmationSectionHeaderText>
+      <SectionHeaderText>FROM</SectionHeaderText>
       <Container paddingHorizontal={3} marginTop={4}>
         <Container flexDirection="row">
           <ContactAvatar
@@ -107,9 +105,7 @@ const TransferSection = ({ data }: { data: TransferDecodedDataType }) => {
 
   return (
     <Container width="100%">
-      <TransactionConfirmationSectionHeaderText>
-        TRANSER THIS CARD
-      </TransactionConfirmationSectionHeaderText>
+      <SectionHeaderText>TRANSER THIS CARD</SectionHeaderText>
       <Container paddingHorizontal={3} marginTop={4}>
         <Container flexDirection="row">
           <Icon name="prepaid-card" />
@@ -140,9 +136,7 @@ const TransferSection = ({ data }: { data: TransferDecodedDataType }) => {
 const ToSection = ({ data }: { data: TransferDecodedDataType }) => {
   return (
     <Container width="100%">
-      <TransactionConfirmationSectionHeaderText>
-        TO
-      </TransactionConfirmationSectionHeaderText>
+      <SectionHeaderText>TO</SectionHeaderText>
       <Container paddingHorizontal={3} marginTop={4}>
         <Container flexDirection="row">
           <Icon name="user" />
