+++ cardstack/src/components/TransactionConfirmationSheet/displays/components/sections/PrepaidCardTransactionSection.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import MiniPrepaidCard from '../PrepaidCard/MiniPrepaidCard';
-import TransactionListItem from './components/TransactionListItem';
+import MiniPrepaidCard from '../../../../PrepaidCard/MiniPrepaidCard';
+import TransactionListItem from '../TransactionListItem';
 import { Container, Text } from '@cardstack/components';
 import {
   useNativeCurrencyAndConversionRates,
